requirements.txt
pydantic>=2.8.0
requests>=2.32.0
gpxpy>=1.6.2
python-dotenv>=1.0.1